fix(kvstore): move Locking
