+++ packages/shared/lib/index.ts
@@ -34,9 +34,6 @@ export * from './services/proxy/request.js';
 
 export * as oauth2Client from './clients/oauth2.client.js';
 
-export * from './utils/lock/locking.js';
-export * from './clients/locking.js';
-
 export * from './models/index.js';
 
 export * from './utils/utils.js';
