+++ packages/shared/lib/services/connection.service.ts
@@ -72,10 +72,10 @@ import type { Orchestrator } from '../clients/orchestrator.js';
 import { SlackService } from './notification/slack.service.js';
 import { getProvider } from './providers.js';
 import { v4 as uuidv4 } from 'uuid';
-import { locking } from '../clients/locking.js';
-import type { Lock, Locking } from '../utils/lock/locking.js';
 import { generateWsseSignature } from '../signatures/wsse.signature.js';
 import tracer from 'dd-trace';
+import { getLocking } from '@nangohq/kvstore';
+import type { Lock } from '@nangohq/kvstore';
 
 const logger = getLogger('Connection');
 const ACTIVE_LOG_TABLE = dbNamespace + 'active_logs';
@@ -85,12 +85,6 @@ const DEFAULT_BILL_EXPIRES_AT_MS = 35 * 60 * 1000; //This ensures we have an exp
 type KeyValuePairs = Record<string, string | boolean>;
 
 class ConnectionService {
-    private locking: Locking;
-
-    constructor(locking: Locking) {
-        this.locking = locking;
-    }
-
     public generateConnectionId(): string {
         return uuidv4();
     }
@@ -1211,6 +1205,7 @@ class ConnectionService {
         }>
     > {
         const providerConfigKey = providerConfig.unique_key;
+        const locking = await getLocking();
 
         // fetch connection and return credentials if they are fresh
         const getConnectionAndFreshCredentials = async (): Promise<{
@@ -1271,7 +1266,7 @@ class ConnectionService {
             let connectionToRefresh: DBConnectionDecrypted;
             try {
                 const lockKey = `lock:refresh:${environment_id}:${providerConfigKey}:${connectionId}`;
-                lock = await this.locking.tryAcquire(lockKey, ttlInMs, acquisitionTimeoutMs);
+                lock = await locking.tryAcquire(lockKey, ttlInMs, acquisitionTimeoutMs);
                 // Another refresh was running so we check if the credentials were refreshed
                 // If yes, we return the new credentials
                 // If not, we proceed with the refresh
@@ -1339,7 +1334,7 @@ class ConnectionService {
             return { success: false, error, response: null };
         } finally {
             if (lock) {
-                await this.locking.release(lock);
+                await locking.release(lock);
             }
         }
     }
@@ -2088,4 +2083,4 @@ class ConnectionService {
     }
 }
 
-export default new ConnectionService(locking);
+export default new ConnectionService();
