--- packages/shared/lib/utils/kvstore/KVStore.ts
@@ -1,6 +1,0 @@
-export interface KVStore {
-    set(key: string, value: string, canOverride?: boolean, ttlInMs?: number): Promise<void>;
-    get(key: string): Promise<string | null>;
-    delete(key: string): Promise<void>;
-    exists(key: string): Promise<boolean>;
-}
