+++ packages/shared/package.json
@@ -22,6 +22,7 @@
         "@datadog/datadog-api-client": "1.26.0",
         "@hapi/boom": "^10.0.1",
         "@nangohq/database": "file:../database",
+        "@nangohq/kvstore": "file:../kvstore",
         "@nangohq/nango-yaml": "file:../nango-yaml",
         "@nangohq/providers": "file:../providers",
         "@nangohq/utils": "file:../utils",
@@ -42,7 +43,6 @@
         "parse-link-header": "^2.0.0",
         "posthog-node": "^3.6.3",
         "qs": "^6.11.2",
-        "redis": "4.6.13",
         "semver": "^7.5.4",
         "simple-oauth2": "5.1.0",
         "uuid": "^9.0.0"
